image_generation/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
beautifulsoup4==4.12.2
blis==0.7.11
cachetools==5.5.1
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
decorator==5.1.1
en-core-sci-sm @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.0/en_core_sci_sm-0.5.0.tar.gz
flatbuffers==25.2.10
fonttools==4.56.0
gast==0.6.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.12.1
html5lib==1.1
idna==2.8
imageio==2.27.0
imgaug==0.4.0
imutils==0.5.4
Jinja2==3.1.5
joblib==1.3.2
keras==2.14.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
marisa-trie==1.2.1
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
ml-dtypes==0.2.0
murmurhash==1.0.12
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.6.0.66
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.2
pathlib_abc==0.1.1
pathy==0.11.0
pillow==10.3.0
preshed==3.0.9
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==1.8.2
pyparsing==3.2.1
pypng==0.20220715.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyWavelets==1.8.0
PyYAML==6.0.2
qrcode==7.4.2
requests==2.21.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.4.2
scipy==1.15.1
seaborn==0.12.2
shapely==2.0.7
six==1.17.0
smart-open==6.4.0
soundfile==0.13.1
soupsieve==2.6
spacy==3.2.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
thinc==8.0.17
threadpoolctl==3.5.0
tifffile==2025.1.10
tqdm==4.67.1
typer==0.4.2
typing_extensions==4.5.0
tzdata==2025.1
urllib3==1.24.3
validators==0.18.2
wasabi==0.10.1
webencodings==0.5.1
Werkzeug==3.1.3
wfdb==4.1.2
wrapt==1.14.1
